+++ examples/histogram.jl
@@ -1,7 +1,7 @@
 # INCLUDE ROCM
 using KernelAbstractions, Test
 using KernelAbstractions: @atomic, @atomicswap, @atomicreplace
-import KernelAbstractions.KernelInterface as KI
+import KernelInterface as KI
 
 include(joinpath(dirname(pathof(KernelAbstractions)), "../examples/utils.jl")) # Load backend
